+++ client/src/components/shared/navpills/index.js
@@ -14,5 +14,8 @@ const Navpills = () =>
         <li className={window.location.pathname === "/applicant" ? "active" : ""}>
             <Link to="/applicant">Applicants</Link>
         </li>
+        <li className={window.location.pathname === "/postjob" ? "active" : ""}>
+            <Link to="/postjob">Post Jobs</Link>
+        </li>
     </ul>;
 export default Navpills;
